vllm/requirements.txt
modal==0.65.0
openai==1.76.0
requests==2.32.3
python-dotenv==1.0.1
